Quase

+++ Documentos/excluir.php
@@ -8,12 +8,22 @@ $Imagem = mysqli_fetch_assoc($sqlImagem);
 $sql2 = "DELETE FROM tabela_assoc WHERE id_doc='$id'";
 $resultado2 = mysqli_query($conexao, $sql2);
 $excluindoFav = "DELETE FROM favoritos WHERE id_documento = '$id'";
-$resultSet = mysqli_query($conexao, $excluindoFav);
-$sql = "DELETE FROM documentos WHERE idDoc='$id'";
-$resultado = mysqli_query($conexao, $sql);
-mysqli_close($conexao);
-if ($resultado && $resultado2) {
-    unlink("../upload/" . $Imagem['imagem']);
-    header("Location:../Inicio/listaDocs.php");
+$sql = "SELECT * FROM comentarios WHERE idDocumento = '$id'";
+$query = mysqli_query($conexao, $sql);
+$comentarios = mysqli_fetch_all($query, MYSQLI_ASSOC);
+foreach ($comentarios as $comentario) {
+    $sql = "DELETE FROM respostas WHERE idComentario = '$comentario[id]'";
+    $query = mysqli_query($conexao, $sql);
+}
+$sql = "DELETE FROM comentarios WHERE idDocumento = '$id'";
+$query = mysqli_query($conexao, $sql);
+if ($query) {
+    $resultSet = mysqli_query($conexao, $excluindoFav);
+    $sql = "DELETE FROM documentos WHERE idDoc='$id'";
+    $resultado = mysqli_query($conexao, $sql);
+    if ($resultado && $resultado2) {
+        unlink("../upload/" . $Imagem['imagem']);
+        header("Location:../Inicio/listaDocs.php");
+    }
 }
 ?>
